added initial modules and some tests (comments)
Just exploring the interface a bit at present.  The only test is actually just code to help in this exploration.  Committing now because I just realised that the SessionT class doesn’t need the s parameter as this is not variable: it’s always known.  We want to go with a “driver” approach (structure of functions) not a type class so there should only be one possible type to fill this role