Converted to use a command strategy (comments)
The strategy now is that the table is responsible for defining it’s “serialisable” instance but the session converts that to a set of commands that the driver turns into the appropriate action for what ever back end it is driving.

I’m still exploring this to see if it’s the way to go.