Added new ColumnMetaData class with type families (comments)
This is to clean up the code a little.  There have already been advantages in that the lens held in the tripple is locked into exactly one type, and therefor operation (there may be a way to overcome this) where as the lens function in the ColumnMetaData class is fully polymorphic and can be used for get, modify and set.