Removed Serialisation (comments)
The Serialisation module (which I never liked) has been completely removed.  All the work it was doing is now handled via type classes and type families.  I’m not completely satisfied but this is looking to have more potential and involves less rewriting of data.

I also added the relevant instances for all the test records and test code for all of them.