completed implementation of fetch and fetchAll (comments)
Things have worked out better than I had hoped.  The syntax doesn’t look bad and inference is working perfectly.

The problem is that I’m painted into a corner with this.  You can only search on a single field, as there is no way to introduce and/or with this layout.  I’ll need to back out at least one iteration, back to the stage where fetch is explicitly doing something to the expression it receives.  This should give us the ability to add type safe and/or operations.  And it should clean the syntax up to be something like

fetch $ expression

again, as it should be.